Big update to game engine
engine.js now contains the Game Engine abstraction. core.js contains the
Vilify core. This is where direct implementations of Entities are and
game settings are set. main.js is now where assets are loaded and the
game is initiated.

Entities have been updated so that their x and y positions are the
center of the entitity, not the top-left corner.

CSS is now used to set the canvas's border so that we don't need added
complexity in calculating the position of objects on the canvas.