Update objects.json
Added a new name for the hero carrier, then put in some experimental data values for the boss, tank and new Hero carrier. If 1=amount of damage a BASE tower can do in 1 second, 1 range=1 tile, and rate = firerate, we would all be screwed fighting the boss if not for movement speed.
To Do: (Get Logan to confirm what unit one is relative to, correct the values (if needed) and add more values for different heros!